+++ .storybook/version-picker/constants.js
@@ -1,0 +1,2 @@
+export const ADDON_ID = "storybook/version-picker";
+export const TOOL_ID = `${ADDON_ID}/version-picker`;