+++ src/style/themes/mint/mint-theme.spec.js
@@ -6,4 +6,8 @@ describe("mintTheme", () => {
   it("contains the base theme", () => {
     assertIsSubset(baseTheme, mintTheme);
   });
+
+  it("has no accidentally modified tokens in compatibility section", () => {
+    expect(mintTheme.compatibility).toMatchSnapshot();
+  });
 });